Update projections for QAnyStringView, QByteArray, QByteArrayView, and structured data types (#118)
* config-libraries.go: Update export for library containers

* emitcabi.go: Update projections for QAnyStringView, QByteArray, and QByteArrayView

* emitzig.go: Update projections for QAnyStringView, QByteArray, QByteArrayView, and structured data types

* Update bindings and wrappers

* Closes #117

+++ include/extras-karchive/libkzipfileentry.zig
@@ -339,7 +339,7 @@ pub const kzipfileentry = struct {
     ///
     /// ` self: QtC.KZipFileEntry `
     ///
-    pub fn Permissions(self: ?*anyopaque) u16 {
+    pub fn Permissions(self: ?*anyopaque) u32 {
         return @intCast(qtc.KArchiveEntry_Permissions(@ptrCast(self)));
     }
